+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Changes to Calva.
 
 ## [Unreleased]
+
+## [2.0.220] - 2021-10-20
 - [Add Paredit select/kill right functionality](https://github.com/BetterThanTomorrow/calva/issues/1024)
 
 ## [2.0.219] - 2021-10-19